requirements.txt
absl-py==2.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
astor==0.8.1
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.13.4
bitarray==3.4.3
blinker==1.9.0
boto3==1.38.42
botocore==1.38.42
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
choreographer==1.0.10
ckzg==2.1.1
click==8.2.1
contourpy==1.3.3
courlan==1.3.2
cryptography==45.0.4
cycler==0.12.1
cytoolz==1.0.1
dash==2.18.2
dash-bootstrap-components==1.7.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash_daq==0.5.0
datasets==4.0.0
dateparser==1.2.2
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
channels==4.1.0
uvicorn==0.32.0
decorator==5.2.1
dill==0.3.8
distro==1.9.0
docutils==0.21.2
dtale==3.18.2
et_xmlfile==2.0.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
evdev==1.9.2
exceptiongroup==1.3.0
executing==2.2.1
feedparser==6.0.11
filelock==3.18.0
filetype==1.2.0
Flask==3.0.3
Flask-Compress==1.18
flatbuffers==25.2.10
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.3.0
future==1.0.0
gast==0.6.0
google-auth==2.40.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gql==3.5.3
graphql-core==3.2.6
grpcio==1.73.0
h11==0.16.0
h5py==3.14.0
hexbytes==1.3.1
hf-xet==1.1.5
htmldate==1.9.3
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
ipython==9.5.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.2
jusText==3.0.2
kaleido==1.0.0
keras==3.5.0
Kivy==2.3.1
Kivy-Garden==0.1.5
kiwisolver==1.4.9
libclang==18.1.1
logistro==1.1.0
lxml==5.4.0
lxml_html_clean==0.4.2
lz4==4.4.4
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
missingno==0.5.2
ml_dtypes==0.5.1
multidict==6.5.0
multiprocess==0.70.16
namex==0.1.0
narwhals==2.3.0
nest-asyncio==1.6.0
networkx==3.5
numpy==1.26.4
oauthlib==3.3.1
openai==1.91.0
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.16.0
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pandasgui==0.2.15
parsimonious==0.10.0
parso==0.8.5
patsy==1.0.1
peewee==3.18.2
pexpect==4.9.0
pillow==11.3.0
pkginfo==1.12.1.2
plotly==6.3.0
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==4.25.8
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pynput==1.8.1
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.17
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-xlib==0.33
pytz==2025.2
pyunormalize==16.0.0
PyYAML==6.0.2
pyzstd==0.17.0
qtstylish==0.1.5
regex==2024.11.6
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retrying==1.4.2
rich==14.0.0
rlp==4.1.0
rsa==4.9.1
s3transfer==0.13.0
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.15.3
seaborn==0.13.2
setuptools==80.9.0
sgmllib3k==1.0.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlite-web==0.6.4
sqlite_bro==0.13.1
squarify==0.4.4
stack-data==0.6.3
statsmodels==0.14.5
strsimpy==0.2.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
tf_keras==2.19.0
threadpoolctl==3.6.0
tld==0.13.1
tokenizers==0.21.2
toolz==1.0.0
tqdm==4.67.1
trafilatura==2.0.0
traitlets==5.14.3
transformers==4.52.4
types-beautifulsoup4==4.12.0.20250516
types-html5lib==1.1.11.20250516
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
wcwidth==0.2.13
web3==7.13.0
websockets==11.0.3
Werkzeug==3.0.6
wheel==0.45.1
wordcloud==1.9.4
wrapt==1.14.1
xarray==2025.8.0
xlrd==2.0.2
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
